Consolidate Memo code
