+++ lib/idl/StdinDispatcher.es6.js
@@ -18,7 +18,7 @@
 
 const stdin = require('process').stdin;
 
-const Memo = require('../Memo.es6.js');
+const Memo = require('./Memo.es6.js');
 
 class StdinDispatcher extends Memo {
   init(opts) {
