+++ lib/memo/ArrayMemo.es6.js
