+++ lib/memo/FileReaderMemo.es6.js
@@ -16,20 +16,26 @@
  */
 'use strict';
 
-const Memo = require('../Memo.es6.js');
+const fs = require('fs');
 
-class URLReaderMemo extends Memo {
-  f (str) {
-    this.logger.log(`Scraping URLs from string of length ${str.length}`);
-    const re = /https?:\/\/[^/]+(\/[^?#, \n]*)?(\?[^#, \n]*)?/g;
-    let results = [];
-    let result;
-    while (result = re.exec(str)) {
-      this.logger.log(`Found URL: ${result[0]}`);
-      results.push(result[0]);
-    }
-    return Promise.resolve(results);
+const Memo = require('./Memo.es6.js');
+
+class FileReaderMemo extends Memo {
+  f (path) {
+    this.logger.log(`Reading from ${path}`);
+    return new Promise((resolve, reject) => {
+      fs.readFile(path, (err, data) => {
+        if (err) {
+          this.logger.error(`Error reading from ${path}`);
+          reject(err);
+        } else {
+          const str = data.toString();
+          this.logger.log(`Read ${str.length} characters from ${path}`);
+          resolve(data.toString());
+        }
+      });
+    });
   }
 }
 
-module.exports = URLReaderMemo;
+module.exports = FileReaderMemo;
