+++ lib/memo/Memo.es6.js
@@ -16,9 +16,9 @@
  */
 'use strict';
 
-const Base = require('./Base.es6.js');
-const MCache = require('./cache/MCache.es6.js');
-const PromiseCache = require('./cache/PromiseCache.es6.js');
+const Base = require('../Base.es6.js');
+const MCache = require('../cache/MCache.es6.js');
+const PromiseCache = require('../cache/PromiseCache.es6.js');
 
 function hashCode(str) {
   var hash = 0;
