+++ lib/memo/REMatchMemo.es6.js
@@ -16,7 +16,7 @@
  */
 'use strict';
 
-const Memo = require('../Memo.es6.js');
+const Memo = require('./Memo.es6.js');
 
 class REMatchMemo extends Memo {
   createRE() {
